+++ lib/bundler/audit/database.rb
@@ -82,6 +82,9 @@ module Bundler
       #
       # Updates the ruby-advisory-db.
       #
+      # @param [Boolean, quiet]
+      #   Specify whether `git` should be `--quiet`.
+      #
       # @return [Boolean, nil]
       #   Specifies whether the update was successful.
       #   A `nil` indicates no update was performed.
@@ -91,15 +94,22 @@ module Bundler
       #
       # @since 0.3.0
       #
-      def self.update!
+      def self.update!(options={})
+        raise "Invalid option(s)" unless (options.keys - [:quiet]).empty?
         if File.directory?(USER_PATH)
           if File.directory?(File.join(USER_PATH, ".git"))
             Dir.chdir(USER_PATH) do
-              system 'git', 'pull', 'origin', 'master'
+              command = %w(git pull)
+              command << '--quiet' if options[:quiet]
+              command << 'origin' << 'master'
+              system *command
             end
           end
         else
-          system 'git', 'clone', URL, USER_PATH
+          command = %w(git clone)
+          command << '--quiet' if options[:quiet]
+          command << URL << USER_PATH
+          system *command
         end
       end