+++ ChangeLog.md
@@ -1,3 +1,23 @@
+### Unreleased
+
+* Added `--quiet` option to `check` and `update` commands.
+
+### 0.5.0 / 2016-02-28
+
+* Added {Bundler::Audit::Task}.
+* Added {Bundler::Audit::Advisory#date}.
+* Added {Bundler::Audit::Advisory#cve_id}.
+* Added {Bundler::Audit::Advisory#osvdb_id}.
+* Allow insecure gem sources (`http://` and `git://`), if they are hosted on a
+  private network.
+
+#### CLI
+
+* Added the `--update` option to `bundle-audit check`.
+* `bundle-audit update` now returns a non-zero exit status on error.
+* `bundle-audit update` only updates `~/.local/share/ruby-advisory-db`, if it is a git
+  repository.
+
 ### 0.4.0 / 2015-06-30
 
 * Require ruby >= 1.9.3 due to i18n gem deprecating < 1.9.3.